cloud-server/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
pyjwt>=2.8.0
